YARN-2229. Changed the integer field of ContainerId to be long type. Contributed by Tsuyoshi OZAWA (cherry picked from commit 3122daa80261b466e309e88d88d1e2c030525e3f)
Conflicts:
	hadoop-mapreduce-project/hadoop-mapreduce-client/hadoop-mapreduce-client-app/src/main/java/org/apache/hadoop/mapreduce/v2/app/rm/preemption/CheckpointAMPreemptionPolicy.java

(cherry picked from commit f21a5236e5716f2f78bccb45b1428a679780f05a)

Conflicts:
	hadoop-yarn-project/CHANGES.txt

+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-resourcemanager/src/main/java/org/apache/hadoop/yarn/server/resourcemanager/RMContext.java
@@ -102,5 +102,5 @@ public interface RMContext {
 
   boolean isWorkPreservingRecoveryEnabled();
   
-  int getEpoch();
+  long getEpoch();
 }
